requirements.txt
flask-restplus==0.10.1
Flask-SQLAlchemy==2.2
sqlalchemy-migrate==0.11.0